requirements.txt
torchtyping
typeguard
git+https://github.com/finetuneanon/transformers.git#egg=transformers
gdown 
tqdm
timm
git+https://github.com/openai/CLIP.git
deepspeed
wandb
transformers==4.24.0

